Upgrade Trustfinder.framework
