--- TrustDefender.framework/Modules
@@ -1,1 +1,0 @@
-Versions/Current/Modules
