+++ TrustDefender.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module TrustDefender {
+  umbrella header "TrustDefender.h"
+
+  export *
+  module * { export * }
+}
