--- TrustDefender.framework/Versions/Current
@@ -1,1 +1,0 @@
-A
