+++ TrustDefender.framework/_CodeSignature/CodeResources
@@ -1,0 +1,181 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Headers/THMProfileHandle.h</key>
+		<data>
+		ZaaKC5TEGbpDWv3D5Aa8Qa7CDcg=
+		</data>
+		<key>Headers/THMStatusCode.h</key>
+		<data>
+		agIKn0xzjTAgxwnnSjXv94GXqts=
+		</data>
+		<key>Headers/TrustDefender.h</key>
+		<data>
+		LyspG5pXCOBvnLEe0mxccSYhprA=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		hd7shtEVyhMbM75kH+Y6yi3DNig=
+		</data>
+		<key>Modules/module.modulemap</key>
+		<data>
+		OuBwU0zvA6T4lG+oMfjAiv3uL8s=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/THMProfileHandle.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			ZaaKC5TEGbpDWv3D5Aa8Qa7CDcg=
+			</data>
+			<key>hash2</key>
+			<data>
+			7uai/Hyx/2Kh2AS+d73p6/Scg2is4aLgpIFJN4dgsiw=
+			</data>
+		</dict>
+		<key>Headers/THMStatusCode.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			agIKn0xzjTAgxwnnSjXv94GXqts=
+			</data>
+			<key>hash2</key>
+			<data>
+			E0HnQtci9wREaLmKnvWzCaUxS3q+a2bG3x7NciScknY=
+			</data>
+		</dict>
+		<key>Headers/TrustDefender.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			LyspG5pXCOBvnLEe0mxccSYhprA=
+			</data>
+			<key>hash2</key>
+			<data>
+			y7ECSWu/AhfBqHbZwkotbdx5/G37dkh6iZyM1QovP8I=
+			</data>
+		</dict>
+		<key>Modules/module.modulemap</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			OuBwU0zvA6T4lG+oMfjAiv3uL8s=
+			</data>
+			<key>hash2</key>
+			<data>
+			95XJp+bJVGVWGK8ULLGIb1/FCbozT5KnYdw+T1L7RF0=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^(Frameworks|SharedFrameworks|PlugIns|Plug-ins|XPCServices|Helpers|MacOS|Library/(Automator|Spotlight|LoginItems))/</key>
+		<dict>
+			<key>nested</key>
+			<true/>
+			<key>weight</key>
+			<real>10</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^[^/]+$</key>
+		<dict>
+			<key>nested</key>
+			<true/>
+			<key>weight</key>
+			<real>10</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
